requirements.txt
mcp>=0.1.0
requests>=2.31.0
requests-oauthlib>=1.3.1
python-dotenv>=1.0.0


